+++ packages/pandas-gbq/docs/source/changelog.rst
@@ -10,9 +10,15 @@ Changelog
   argument to limit the number of rows in the results DataFrame. Set
   ``max_results`` to 0 to ignore query outputs, such as for DML or DDL
   queries. (:issue:`102`)
+- Add ``progress_bar_type`` argument to :func:`~pandas_gbq.read_gbq()`. Use
+  this argument to display a progress bar when downloading data.
+  (:issue:`182`)
 
-- Add ``progress_bar_type`` argument to :func:`~pandas_gbq.read_gbq()`. Use this
-  argument to display a progress bar when downloading data. (:issue:`182`)
+Dependency updates
+~~~~~~~~~~~~~~~~~~
+
+- Update the minimum version of ``google-cloud-bigquery`` to 1.11.1.
+  (:issue:`296`)
 
 Documentation
 ~~~~~~~~~~~~~